fix: seperate individual editor script from collab editor script. integrate nodemon into concurrently
