+++ package.json
@@ -3,8 +3,8 @@
   "version": "1.0.0",
   "main": "index.js",
   "scripts": {
-    "dev": "concurrently \"tailwindcss -i ./src/styles/style.css -o ./static/styles/style.css --watch\" \"esbuild ./src/scripts/app.js --bundle --outfile=./static/scripts/app.js --watch\" \"node --watch-path=./src/scripts/editors.js ./esbuildEditors.js\"",
-    "dev2": "nodemon --watch src/scripts/editors.js --exec \"node ./esbuildEditors.js\""
+    "watch:editor": "nodemon --watch src/scripts/editors/collab.js --watch src/scripts/editors/individual.js --exec \"node ./esbuildEditors.js\"",
+    "dev": "concurrently \"tailwindcss -i ./src/styles/style.css -o ./static/styles/style.css --watch\" \"esbuild ./src/scripts/app.js --bundle --outfile=./static/scripts/app.js --watch\" \"npm run watch:editor\""
   },
   "keywords": [],
   "author": "",
